chore: move cursor component
