+++ internal/view/components/cursor/cursor.go
