+++ internal/view/components/cursor/cursor_test.go
