+++ internal/view/components/deckview/deckview.go
@@ -2,8 +2,8 @@ package deckview
 
 import (
 	"2sp/internal/config"
+	"2sp/internal/view/components/cursor"
 	"2sp/internal/view/components/voteview"
-	"2sp/internal/view/cursor"
 	"2sp/internal/view/messages"
 	"2sp/pkg/protocol"
 	tea "github.com/charmbracelet/bubbletea"
