+++ internal/view/components/issuesview/issues_view.go
@@ -2,8 +2,8 @@ package issuesview
 
 import (
 	"2sp/internal/config"
+	"2sp/internal/view/components/cursor"
 	"2sp/internal/view/components/voteview"
-	"2sp/internal/view/cursor"
 	"2sp/internal/view/messages"
 	"2sp/pkg/protocol"
 	"fmt"
